Settle on "data partition" terminology
The document uses the terms "leaf parquet file", "data partition", and "pixel" to describe the same files on disk. In the beginning of the document, as the spatial sorting is described, it makes sense to describe leaf parquet files and pixels, but as the document continues, it settles on "data partition" or "partition", which I think is fine. Changing one of the later sections to match this.